Refactor to remove the necessity to have a separate ElementWrapper type
