+++ src/content/actions/hoverElement.ts
@@ -1,4 +1,4 @@
-import { type ElementWrapper } from "../../typings/ElementWrapper";
+import { type ElementWrapper } from "../wrappers/ElementWrapper";
 import { getHintedWrappers } from "../wrappers/wrappers";
 
 export async function hoverElement(wrappers: ElementWrapper[]) {
